stock_monitoring/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
newsapi-python>=0.2.7
yfinance>=0.2.36
jinja2>=3.1.2
google-cloud-core
google-adk
